+++ Tests/Ontology-InvokeTests.swift
@@ -1,0 +1,147 @@
+//
+//  Ontology-InvokeTests.swift
+//  neovmUtils_Tests
+//
+//  Created by Wyatt Mufson on 10/2/19.
+//  Copyright © 2019 Ryu Blockchain Technologies. All rights reserved.
+//
+
+import XCTest
+import Neoutils
+
+class Ontology_InvokeTests: XCTestCase {
+    var exampleWallet: Wallet = newWallet()
+
+    override func setUp() {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testBuildJoinTransaction() {
+        let contractHash = "6b21a978e40e681c8439e2fb9cb39424920bf3e1"
+        let gid = "G1"
+        let mid = "M1"
+        let entry = 10.0
+        let max : Int = 1
+
+        let target = NVMParameter(type: .Address, value: exampleWallet.address)
+        let gameId = NVMParameter(type: .String, value: gid)
+        let matchId = NVMParameter(type: .String, value: mid)
+        let fee = NVMParameter(type: .Fixed8, value: entry)
+        let mx = NVMParameter(type: .Integer, value: max)
+
+        let args = [target, gameId, matchId, fee, mx]
+
+        let method = "join"
+
+        let gasPrice = 500
+        let gasLimit = 20000
+
+        let tx = buildOntologyInvocationTransaction(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif, payer: exampleWallet.address)
+        XCTAssertNotEqual(tx, "")
+    }
+
+    func testBuildOntologyInvocation() {
+        let contractHash = "c168e0fb1a2bddcd385ad013c2c98358eca5d4dc"
+        let method = "put"
+        let argDict : [[String:Any]] = [["T":"Address", "V":exampleWallet.address], ["T":"String", "V":"Hello!"]]
+
+        do {
+            let data = try JSONSerialization.data(withJSONObject: argDict, options: .prettyPrinted)
+            let args = String(data: data, encoding: String.Encoding.utf8)
+            let gasPrice = 500
+            let gasLimit = 20000
+            let wif = exampleWallet.wif
+            let err = NSErrorPointer(nilLiteral: ())
+            let res = NeoutilsBuildOntologyInvocationTransaction(contractHash, method, args, gasPrice, gasLimit, wif, exampleWallet.address, err)
+            XCTAssertNil(err)
+            XCTAssertNotNil(res)
+        } catch let error {
+            XCTFail("Failed to cast JSON: \(error)")
+        }
+    }
+
+    func testBuildOntologyInvocationHelper() {
+        let contractHash = "c168e0fb1a2bddcd385ad013c2c98358eca5d4dc"
+        let method = "put"
+        let args: [NVMParameter] = [NVMParameter(type: .Address, value: exampleWallet.address), NVMParameter(type: .String, value: "Hello!")]
+        let gasPrice = 500
+        let gasLimit = 20000
+
+        let res = buildOntologyInvocationTransaction(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif, payer: exampleWallet.address)
+        XCTAssertNotEqual("", res)
+
+        let res2 = buildOntologyInvocationTransaction(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif, payer: exampleWallet.address)
+        XCTAssertNotEqual("", res2)
+
+        let res3 = buildOntologyInvocationTransaction(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: "123")
+        XCTAssertEqual("", res3)
+    }
+
+    func testOntologyInvocation() {
+        let contractHash = "c168e0fb1a2bddcd385ad013c2c98358eca5d4dc"
+        let method = "put"
+        let argDict : [[String:Any]] = [["T": "Address", "V": exampleWallet.address], ["T": "String", "V": "Hello!"]]
+
+        do {
+            let data = try JSONSerialization.data(withJSONObject: argDict, options: .prettyPrinted)
+            let args = String(data: data, encoding: String.Encoding.utf8)
+            let gasPrice = 500
+            let gasLimit = 20000
+            let endpoint = "http://polaris2.ont.io:20336"
+            let wif = exampleWallet.wif
+            let err = NSErrorPointer(nilLiteral: ())
+            let res = NeoutilsOntologyInvoke(endpoint, contractHash, method, args, gasPrice, gasLimit, wif, exampleWallet.address, err)
+            XCTAssertNil(err)
+            XCTAssertNotNil(res)
+            print("Response: \(res)")
+        } catch let error {
+            XCTFail("Failed to cast JSON: \(error)")
+        }
+    }
+
+    func testOntologyInvocationHelper() {
+        let contractHash = "c168e0fb1a2bddcd385ad013c2c98358eca5d4dc"
+        let method = "put"
+        let args: [NVMParameter] = [NVMParameter(type: .Address, value: exampleWallet.address), NVMParameter(type: .String, value: "Hello!")]
+        let gasPrice = 500
+        let gasLimit = 20000
+
+        let res = ontologyInvoke(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif, payer: exampleWallet.address)
+        XCTAssertNotEqual("", res)
+
+        let res2 = ontologyInvoke(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif)
+        XCTAssertNotEqual("", res2)
+
+        let res3 = ontologyInvoke(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: "123")
+        XCTAssertEqual("", res3)
+    }
+
+    func testOntologyInvocationRead() {
+        let contractHash = "a29564a30043d50620e4c6be61eda834d0acc48b"
+
+        let methods: [String] = ["getTotal", "getCirculation", "getLocked"]
+        for method in methods {
+            let res = ontologyInvokeRead(contractHash: contractHash, method: method, args: [])
+            let num = res.hexToDecimal()
+            XCTAssertGreaterThan(num, 0)
+            print(num)
+        }
+    }
+
+    func testSendRawTransaction() {
+        let contractHash = "c168e0fb1a2bddcd385ad013c2c98358eca5d4dc"
+        let method = "put"
+        let args: [NVMParameter] = [NVMParameter(type: .Address, value: exampleWallet.address), NVMParameter(type: .String, value: "Hello!")]
+        let gasPrice = 500
+        let gasLimit = 20000
+
+        let res = buildOntologyInvocationTransaction(contractHash: contractHash, method: method, args: args, gasPrice: gasPrice, gasLimit: gasLimit, wif: exampleWallet.wif, payer: exampleWallet.address)
+        let txId = ontologySendRawTransaction(raw: res)
+        XCTAssertNotEqual("", res)
+        XCTAssertNotEqual("", txId)
+    }
+}
